feat: Error handler code isolation from main.py
