+++ src/state_db/main.py
@@ -18,6 +18,7 @@ from state_db.configs import (
     CORS_ORIGINS,
     LOGGING_CONFIG,
 )
+from state_db.configs.exceptions import init_exception_handlers
 from state_db.custom import CustomJSONResponse
 
 logger = logging.getLogger("uvicorn.error")
@@ -63,59 +64,7 @@ async def error_logging_middleware(request: Request, call_next):
     return response
 
 
-# 1. 일반적인 모든 서버 에러 (500)
-@app.exception_handler(Exception)
-async def universal_exception_handler(request: Request, exc: Exception):
-    logger.error(f"🔥 Unexpected Error: {request.method} {request.url.path}")
-    logger.error(traceback.format_exc())
-    return JSONResponse(
-        status_code=500,
-        content={
-            "status": "error",
-            "message": "서버 내부 오류가 발생했습니다.",
-            "detail": str(exc),
-        },
-    )
-
-
-# 2. 의도된 HTTP 에러 (400, 401, 404, 503 등) - 중복 제거 및 통합
-@app.exception_handler(HTTPException)
-async def http_exception_handler(request: Request, exc: HTTPException):
-    logger.error(f"⚠️ HTTP {exc.status_code} Error: {request.method} {request.url.path}")
-    logger.error(f"Detail: {exc.detail}")
-    return JSONResponse(
-        status_code=exc.status_code,
-        content={
-            "status": "error",
-            "message": "요청 처리 중 오류가 발생했습니다.",
-            "detail": exc.detail,
-        },
-    )
-
-
-# 3. 데이터 검증 에러 (422)
-@app.exception_handler(RequestValidationError)
-async def validation_exception_handler(request: Request, exc: RequestValidationError):
-    errors = exc.errors()
-    error_details = []
-    for error in errors:
-        loc = " -> ".join([str(x) for x in error.get("loc", [])])
-        msg = error.get("msg")
-        inp = error.get("input")
-        error_details.append(f"[{loc}] {msg} (Input: {inp})")
-
-    full_message = " | ".join(error_details)
-    logger.error(f"❌ Validation Error: {request.method} {request.url.path}")
-    logger.error(f"Detail: {full_message}")
-
-    return JSONResponse(
-        status_code=422,
-        content={
-            "status": "error",
-            "message": "입력값 검증에 실패했습니다.",
-            "detail": errors,
-        },
-    )
+init_exception_handlers(app)
 
 
 app.add_middleware(
